add tests for library, test int/bs parameters
