+++ test/KadenaBridgeWallet.js
@@ -13,64 +13,6 @@ let creator;
 let owner;
 let oracle;
 
-// web3's `getGasPrice` returns value in wei
-let avgGasPrice = 47000000000; // in wei (average in last year)
-let slowerGasPrice = 29000000000; // in wei
-let veryFastGasPrice = 70000000000; // in wei
-let etherToUSD = 614.82;  // value of 1 ether in USD (Dec 2020)
-
-function makeLeafHash(str) {
-  let inputHex = web3.utils.toHex(str);
-      inputBytes = web3.utils.hexToBytes(inputHex);
-      leafTagBytes = web3.utils.hexToBytes("0x00");
-      bytes = leafTagBytes.concat(inputBytes);
-      hex = web3.utils.bytesToHex(bytes);
-      hash = web3.utils.soliditySha3(hex);
-  return hash;  // in hex string
-}
-
-function makeNodeHash(hexA, hexB) {
-  let nodeTagBytes = web3.utils.hexToBytes("0x01");
-      bytesA = web3.utils.hexToBytes(hexA);
-      bytesB = web3.utils.hexToBytes(hexB);
-      bytes = nodeTagBytes.concat(bytesA.concat(bytesB));
-      hex = web3.utils.bytesToHex(bytes);
-      hash = web3.utils.soliditySha3(hex);
-  return hash;  // in hex string
-}
-
-function genProofN(proofLength) {
-  let subject = makeLeafHash("hello");
-      worldLeafHash = makeLeafHash("world");
-      hashes = Array.from({length: proofLength}, (_, i) => worldLeafHash);
-      sides = Array.from({length: proofLength}, (_, i) => "0x01");
-
-  return {
-    subject,
-    hashes,
-    sides,
-    proofLength
-  };
-};
-
-function getGasCosts(gas, gasPrice, etherToUSD) {
-  let wei = Number(gas * gasPrice);
-      ether = web3.utils.fromWei(wei.toFixed(0), 'ether');
-      usd = ether * etherToUSD;
-
-  console.log("Gas Price is " + gasPrice + " wei");
-  console.log("Gas estimation = " + gas + " units");
-  console.log("Gas cost estimation = " + wei + " wei");
-  console.log("Gas cost estimation = " + ether + " ether");
-  console.log("Gas cost estimation = " + usd + " USD");
-
-  return {
-    gas,
-    wei,
-    ether,
-    usd
-  };
-};
 
 contract ('KadenaBridgeWallet', (accounts) => {
 
@@ -87,181 +29,6 @@ contract ('KadenaBridgeWallet', (accounts) => {
           {from: creator});
         oracle = oracleObj.address;
 	});
-
-  it("Validate simple Merkle proof", async () => {
-      // Create the wallet contract
-      let kadenaBridgeWallet = await KadenaBridgeWallet.new(creator, owner, "someChainwebPublicKey", oracle);
-
-      // Not valid merkle tree. Used to test iteration and hashing logic.
-      let hello = web3.utils.toHex("hello");
-          helloLeafHash = await kadenaBridgeWallet.hashLeaf(hello);
-          worldLeafHash = makeLeafHash("world");
-          expectedRoot = makeNodeHash(helloLeafHash, worldLeafHash);
-
-      assert(helloLeafHash == "0xaa872873635ad305d25327a952b25396b95b3ddfcfd661ab26241a853f70451c",
-            "Leaf hash of `hello` DOES NOT match Haskell calculated hash");
-      assert(worldLeafHash == "0x6fb62750ef421842f3daf5b532cb72ab6530a5e3005b7e76295717fd4d272a66",
-            "Leaf hash of `world` DOES NOT match Haskell calculated hash");
-      assert(expectedRoot == "0xe05caef6cbd3e77acdea8693af934b82b5b6ceeca539601af8deeede7c6455a1",
-            "Expected root DOES NOT Haskell calculated hash");
-
-      let actualRoot = await kadenaBridgeWallet.runMerkleProof(
-            hello,   // subject hex
-            1, // proof path step count
-            [worldLeafHash], // proof path hashes
-            ["0x01"], // proof path sides (adds path proof to right)
-            {from: creator});
-
-      assert(expectedRoot == actualRoot,
-        "Expected and actual root DO NOT match");
-  });
-
-  it("Validate Merkle subject bytes encoding", async () => {
-      // Create the wallet contract
-      let kadenaBridgeWallet = await KadenaBridgeWallet.new(creator, owner, "someChainwebPublicKey", oracle);
-
-      // Encodes request key "VlVm90jiVdlkQ9frX_XtVtYeTj5lTOlcVkj65jsjSUM"
-      // Base64 URL No Padding: "ACsAAABWbFZtOTBqaVZkbGtROWZyWF9YdFZ0WWVUajVsVE9sY1ZrajY1anNqU1VN"
-      let rkBytes = "0x002b000000566c566d39306a6956646c6b51396672585f587456745965546a356c544f6c63566b6a36356a736a53554d";
-          rkParsed = await kadenaBridgeWallet.parseBytes(rkBytes);
-          rkExpectedHex = "0x566c566d39306a6956646c6b51396672585f587456745965546a356c544f6c63566b6a36356a736a53554d";
-
-      console.log(rkParsed[0]);
-      console.log(rkParsed[1].toString());
-      console.log(rkParsed[2]);
-
-      assert(rkExpectedHex == rkParsed[2],
-        "Failed to decode bytes");
-
-  });
-
-  it("Validate actual Merkle proof", async () => {
-      // Create the wallet contract
-      let kadenaBridgeWallet = await KadenaBridgeWallet.new(creator, owner, "someChainwebPublicKey", oracle);
-
-      /* To replicate in Haskell:
-        `{-# LANGUAGE TypeApplications #-}
-         {-# LANGUAGE OverloadedStrings #-}
-
-         import Data.MerkleLog
-         import qualified Data.ByteString as B
-         import Crypto.Hash.Algorithms (Keccak_256)
-         import qualified Data.ByteArray as BA
-         import qualified Data.Memory.Endian as BA
-
-         inputs = ["a", "b", "c"] :: [B.ByteString]
-         inputs' = map InputNode inputs
-
-         -- create tree
-         t = merkleTree @Keccak_256 inputs'
-
-         -- create inclusion proof
-         p = either (error . show) id $ merkleProof (inputs' !! 1) 1 t
-
-         (MerkleProof proofSubj proofObj) = p
-         subj = case (_getMerkleProofSubject proofSubj) of
-            InputNode b -> BA.convertToBase BA.Base16 (merkleLeaf @a b)
-            TreeNode r -> BA.convertToBase BA.Base16 r
-         // "e99905ac9f9583a5737a07d20a7129343f486f5f549b42c05192046188ef5f66"
-
-         BA.convertToBase @BA.Bytes BA.Base16 $ encodeMerkleProofObject proofObj :: BA.Bytes
-         // "000000020000000000000001009722201502e620d70d78ee63045f3493812c206b988cbbe76c28918a7364fdbd014c89fefa814dbe46b640ca2ffb4682a1eaad32985c6604e98cc0a2fd76e49550"
-
-         runMerkleProof p // P2wtbQwvzWd5XqUK8NyFyOLfiDLv48SeNtj-LnG8wHs
-         merkleRoot t // P2wtbQwvzWd5XqUK8NyFyOLfiDLv48SeNtj-LnG8wHs
-         BA.convertToBase @BA.Bytes BA.Base16 $ encodeMerkleRoot (merkleRoot t) :: BA.Bytes
-         // "3f6c2d6d0c2fcd67795ea50af0dc85c8e2df8832efe3c49e36d8fe2e71bcc07b"
-      */
-
-      // Not valid merkle tree. Used to test iteration and hashing logic.
-      let subj = web3.utils.toHex("b");
-          expectedSubjHash = "0xe99905ac9f9583a5737a07d20a7129343f486f5f549b42c05192046188ef5f66";
-          path = [ "0x9722201502e620d70d78ee63045f3493812c206b988cbbe76c28918a7364fdbd"
-                 , "0x4c89fefa814dbe46b640ca2ffb4682a1eaad32985c6604e98cc0a2fd76e49550"];
-          sides = ["0x00", "0x01"];
-          expectedRoot = "0x3f6c2d6d0c2fcd67795ea50af0dc85c8e2df8832efe3c49e36d8fe2e71bcc07b";
-
-      assert(makeLeafHash("b") == expectedSubjHash,
-            "Expected subject hash is leaf hash");
-      let actualSubjHash = await kadenaBridgeWallet.hashLeaf(subj);
-      assert(expectedSubjHash == actualSubjHash,
-            "Expected subject hash matches leaf hash calculated by solidity contract");
-
-      let actualRoot = await kadenaBridgeWallet.runMerkleProof(
-            subj,   // subject in hex
-            2, // proof path step count
-            path, // proof path hashes
-            sides, // proof path sides (adds path proof to right)
-            {from: creator});
-
-      assert(expectedRoot == actualRoot,
-        "Expected and actual root DO NOT match");
-  });
-
-  it("Gas costs for proof SIZE = 10 and AVERAGE gas price", async () => {
-
-      let gasPrice = avgGasPrice;  // in wei
-          etherToUSD = 614.82;  // value in USD of 1 ether
-
-      let kadenaBridgeWallet = await KadenaBridgeWallet.new(
-            creator, owner, "someChainwebPublicKey", oracle
-          );
-
-      let proof = genProofN(10);
-          gasAmount = await kadenaBridgeWallet.runMerkleProof.estimateGas(
-            proof.subject,   // subject hash
-            proof.proofLength, // proof path step count
-            proof.hashes, // proof path hashes
-            proof.sides, // proof path sides (adds path proof to right)
-            {from: creator});
-          gas = Number(gasAmount);
-
-      console.log("Proof length = " + proof.proofLength);
-      let results = getGasCosts(gas, gasPrice, etherToUSD);
-  });
-
-  it("Gas costs for proof SIZE = 100 and AVERAGE gas price", async () => {
-
-      let gasPrice = avgGasPrice;  // in wei
-
-      let kadenaBridgeWallet = await KadenaBridgeWallet.new(
-            creator, owner, "someChainwebPublicKey", oracle
-          );
-
-      let proof = genProofN(100);
-          gasAmount = await kadenaBridgeWallet.runMerkleProof.estimateGas(
-            proof.subject,   // subject hash
-            proof.proofLength, // proof path step count
-            proof.hashes, // proof path hashes
-            proof.sides, // proof path sides (adds path proof to right)
-            {from: creator});
-          gas = Number(gasAmount);
-
-      console.log("Proof length = " + proof.proofLength);
-      let results = getGasCosts(gas, gasPrice, etherToUSD);
-  });
-
-  it("Gas costs for proof SIZE = 500 and AVERAGE gas price", async () => {
-
-      let gasPrice = avgGasPrice;  // in wei
-          etherToUSD = 614.82;  // value in USD of 1 ether
-
-      let kadenaBridgeWallet = await KadenaBridgeWallet.new(
-            creator, owner, "someChainwebPublicKey", oracle
-          );
-
-      let proof = genProofN(500);
-          gasAmount = await kadenaBridgeWallet.runMerkleProof.estimateGas(
-            proof.subject,   // subject hash
-            proof.proofLength, // proof path step count
-            proof.hashes, // proof path hashes
-            proof.sides, // proof path sides (adds path proof to right)
-            {from: creator});
-          gas = Number(gasAmount);
-
-      console.log("Proof length = " + proof.proofLength);
-      let results = getGasCosts(gas, gasPrice, etherToUSD);
-  });
 
   it("Cannot send ETH to bridge wallet via /send", async () => {
       let validProof = "someDummyValidProof";
